Add Linux Docker build and fix cross-platform format warnings
Docker:
- Add docker/Dockerfile (Ubuntu 22.04 + clang + SDL2 + GTK3)
- Add .dockerignore to exclude build artifacts from context

Linux build fix:
- debuggerDesktop.cpp: Replace SetItemTooltipV(fmt, nullptr) with
  SetItemTooltip(fmt) — nullptr cannot convert to va_list on Linux

Cross-platform format specifier casts (int64_t/uint64_t/frame_time_t
are long/unsigned long on Linux but long long on macOS):
- hardfile_host.cpp: Cast 9 format args to (long long)/(unsigned long long)
- mman.cpp: Cast uint64_t args to (unsigned long long)
- hardfile.cpp: Cast virtsize args to (unsigned long long)
- vm.cpp: Cast uae_u64/uintptr_t args to (unsigned long long)
- zfile.cpp: Cast virtsize arg to (unsigned long long)

vAmiga:
- va_server_thread.cpp: Fix visible area cropping to center 754px window
  in 840px active area (matching UAE scaling) and support NTSC height

+++ libs/uae_lib/hardfile.cpp
@@ -747,7 +747,7 @@ int hdf_open (struct hardfiledata *hfd, const TCHAR *pname)
 	}
 	write_log (_T("HDF is VHD %s image, virtual size=%lldK (%llx %lld)\n"),
 		hfd->hfd_type == HFD_VHD_FIXED ? _T("fixed") : _T("dynamic"),
-		hfd->virtsize / 1024, hfd->virtsize, hfd->virtsize);
+		(unsigned long long)(hfd->virtsize / 1024), (unsigned long long)hfd->virtsize, (unsigned long long)hfd->virtsize);
 	hdf_init_cache (hfd);
 	return 1;
 nonvhd:
